+++ src/engine/input/RunOnKeyUpAttribute.cs
@@ -34,7 +34,7 @@ public class RunOnKeyUpAttribute : RunOnKeyAttribute
         return false;
     }
 
-    public override void OnProcess(float delta)
+    public override void OnProcess(double delta)
     {
     }
 }
